FastAPIProject/requirements.txt
matplotlib==3.10.8
numpy==2.4.2
Pillow==12.1.1
torch==2.10.0
torchvision==0.25.0
fastapi
python-multipart
